requirements/prod.txt
# Everything needed in production

#setuptools==18.5
#wheel==0.26.0

# Flask
Flask>=0.11
MarkupSafe>=0.23
Werkzeug>=0.11.10
Jinja2>=2.8
itsdangerous>=0.24
flask_mail>=0.9.1
Flask-JWT-Extended>=0.0.8

# Database
Flask-SQLAlchemy>=2.1
SQLAlchemy>=1.0.13
psycopg2>=2.6.1

# Migrations
Flask-Migrate>=1.8.0

# Forms
Flask-WTF>=0.12
WTForms>=2.1

# Deployment
gunicorn>=19.1.1
gevent>=0.4.7
tornado>=4.2.2

# Auth
Flask-Login>=0.3.2
Flask-Bcrypt>=0.7.1

# Caching
Flask-Cache>=0.13.1
Flask-Compress>=1.3.2

# Debug toolbar
Flask-DebugToolbar>=0.10.0

# Flask Rest Api
flask-restful>=0.3.5

# Text
Twilio>=5.6.0
# Cryptography Bcrypt dependency
cryptography>=1.3.2

# Mapbox
mapbox>=0.8.0

#Stripe
stripe>=1.35.0

# Json Web tokens
Pyjwt>=1.4.0

# Marshmallow
marshmallow>=2.7.3

#Celery
celery>=3.1.23

#Flask-cors
flask-cors>=3.0.0



#Newest versions
#Flask
#MarkupSafe
#Werkzeug
#Jinja2
#itsdangerous
#flask_mail
#Flask-SQLAlchemy
#psycopg2
#Flask-Migrate
#Flask-WTF
#Flask-JWT-Extended
#WTForms
#Gunicorn
#Flask-Assets
#cssmin
#jsmin
#Flask-Login
#Flask-Bcrypt
#Flask-Cache
#Flask-DebugToolbar
#flask-restful
#flask-compress
#cryptography
#mapbox
#stripe
#Pyjwt
#pygeoip
#marshmallow
#celery
#flask-cors
#gevent
#twilio
